Fix fixture build path and revert fixture.test.js
